test: move unit tests
