+++ test/unit/components/VApp/VApp.spec.js
@@ -41,26 +41,4 @@ test('VApp.js', ({ mount }) => {
     })
     expect(wrapper.vm.$vuetify.dark).toBe(false)
   })
-
-  it('should watch theme', async () => {
-    const wrapper = mount(VApp)
-
-    expect(wrapper.vm.style).toMatchSnapshot()
-    wrapper.vm.$vuetify.theme.primary = '#000'
-    await wrapper.vm.$nextTick()
-    expect(wrapper.vm.style).toMatchSnapshot()
-  })
-
-  it('should set a CSP nonce', async () => {
-    // Delete the old stylesheet first
-    let el = document.getElementById('vuetify-theme-stylesheet')
-    el.parentNode.removeChild(el)
-
-    Vue.prototype.$vuetify.options.cspNonce = 'asdfghjkl'
-    const app = mount(VApp, { attachToDocument: true })
-
-    el = document.getElementById('vuetify-theme-stylesheet')
-    expect(el).toBeTruthy()
-    expect(el.getAttribute('nonce')).toBe('asdfghjkl')
-  })
 })
